Overall refactoring and bug fixes

+++ admin/config/uploads.php
@@ -1,8 +1,10 @@
 <?php
 
-require ('config/setup.php');
+$ds = DIRECTORY_SEPARATOR;
 
-$storeFolder = ROOT . D_UPLOADS;
+require ('../config/link.php');
+
+$storeFolder = '../../uploads';
 $id = $_GET['id'];
 
 $ext = pathinfo($_FILES['file']['name'], PATHINFO_EXTENSION);
@@ -10,25 +12,54 @@ $newname = time();
 $random = rand(100, 999);
 $name = $newname . $random . '.' . $ext;
 
-$query = "SELECT avatar FROM users WHERE id = $id";
-$result = mysqli_query($dbc, $query);
-$old = mysqli_fetch_assoc($result);
 
-$query = "UPDATE users SET avatar = '$name' WHERE id = $id";
-$result = mysqli_query($dbc, $query);
+if ($_GET['type'] == 'avatar') {
 
+    $query = "SELECT avatar FROM users WHERE id = $id";
+    $result = mysqli_query($dbc, $query);
+    $old = mysqli_fetch_assoc($result);
 
-if (!empty($_FILES)) {
+    $query = "UPDATE users SET avatar = '$name' WHERE id = $id";
+    $result = mysqli_query($dbc, $query);
 
-    $tempFile = $_FILES['file']['tmp_name'];
-    $targetPath = dirname(__FILE__) . DS . $storeFolder . DS;
-    $targetFile = $targetPath . $name;
+    if (!empty($_FILES)) {
 
-    move_uploaded_file($tempFile, $targetFile);
+        $tempFile = $_FILES['file']['tmp_name'];
+        $targetPath = dirname(__FILE__) . $ds . $storeFolder . $ds;
+        $targetFile = $targetPath . $name;
 
-    $deleteFile = $targetPath . $old['avatar'];
+        move_uploaded_file($tempFile, $targetFile);
 
-    if (!empty($old['avatar']) && !is_dir($deleteFile)) {
-        unlink($deleteFile);
+        $deleteFile = $targetPath . $old['avatar'];
+
+        if (!empty($old['avatar']) && !is_dir($deleteFile)) {
+            unlink($deleteFile);
+        }
     }
 }
+if ($_GET['type'] == 'image') {
+
+    $query = "SELECT image FROM posts WHERE id = $id";
+    $result = mysqli_query($dbc, $query);
+    $old = mysqli_fetch_assoc($result);
+
+    $query = "UPDATE posts SET image = '$name' WHERE id = $id";
+    $result = mysqli_query($dbc, $query);
+
+    if (!empty($_FILES)) {
+
+        $tempFile = $_FILES['file']['tmp_name'];
+        $targetPath = dirname(__FILE__) . $ds . $storeFolder . $ds;
+        $targetFile = $targetPath . $name;
+
+        move_uploaded_file($tempFile, $targetFile);
+
+        $deleteFile = $targetPath . $old['image'];
+
+        if (!empty($old['image']) && !is_dir($deleteFile)) {
+            unlink($deleteFile);
+        }
+    }
+}
+
+
